upgrade emulator functionality for mts200 testing
