+++ custom_components/meross_lan/emulator/emulator.py
@@ -1,0 +1,196 @@
+from time import time
+from json import (
+    dumps as json_dumps,
+    loads as json_loads,
+)
+from typing import Tuple
+from aiohttp.web import (
+    Request as web_Request,
+    json_response as web_json_response,
+)
+
+from ..merossclient import (
+    build_payload,
+    const as mc,
+    get_namespacekey,
+    get_replykey,
+)
+
+from .descriptor import MerossEmulatorDescriptor
+
+class MerossEmulator:
+
+
+    def __init__(self, descriptor: MerossEmulatorDescriptor, key):
+        self.key = key
+        self.descriptor = descriptor
+        self.p_all_system_time = descriptor.system.get(mc.KEY_TIME)
+        if mc.NS_APPLIANCE_SYSTEM_DNDMODE in descriptor.ability:
+            self.p_dndmode = { mc.KEY_DNDMODE: { mc.KEY_MODE: 0 }}
+
+        print(f"Initialized {descriptor.productname} (model:{descriptor.productmodel})")
+
+
+    async def post_config(self, request: web_Request):
+        jsonrequest = await request.json()
+        header:dict = jsonrequest[mc.KEY_HEADER]
+        payload:dict = jsonrequest[mc.KEY_PAYLOAD]
+        namespace:str = header[mc.KEY_NAMESPACE]
+        method:str = header[mc.KEY_METHOD]
+
+        print(f"Emulator({self.descriptor.uuid}) RX: namespace={namespace} method={method} payload={json_dumps(payload)}")
+        try:
+            if self.p_all_system_time is not None:
+                self.p_all_system_time[mc.KEY_TIMESTAMP] = int(time())
+
+            if namespace not in self.descriptor.ability:
+                raise Exception(f"{namespace} not supported in ability")
+
+            elif get_replykey(header, self.key) is not self.key:
+                method = mc.METHOD_ERROR
+                payload = { mc.KEY_ERROR: { mc.KEY_CODE: mc.ERROR_INVALIDKEY} }
+
+            elif (handler := getattr(self, f"_{method}_{namespace.replace('.', '_')}", None)) is not None:
+                method, payload = handler(header, payload)
+
+            else:
+                method, payload = self._handler_default(method, namespace, payload)
+
+        except Exception as e:
+            method = mc.METHOD_ERROR
+            payload = { mc.KEY_ERROR: { mc.KEY_CODE: -1, "message": str(e)} }
+
+        data = build_payload(namespace, method, payload, self.key, mc.MANUFACTURER, header[mc.KEY_MESSAGEID])
+        print(f"Emulator({self.descriptor.uuid}) TX: namespace={namespace} method={method} payload={json_dumps(payload)}")
+        return web_json_response(data)
+
+
+    def _get_key_state(self, namespace: str) -> Tuple[str, dict]:
+        """
+        general device state is usually carried in NS_ALL into the "digest" key
+        and is also almost regularly keyed by using the camelCase of the last verb
+        in namespace.
+        For some devices not all state is carried there tho, so we'll inspect the
+        GETACK payload for the relevant namespace looking for state there too
+        """
+        n = namespace.split('.')
+        if n[1] != 'Control':
+            raise Exception(f"{namespace} not supported in emulator")
+
+        key = get_namespacekey(namespace)
+        p_digest = self.descriptor.digest
+        if len(n) == 4:
+            # 4 parts namespaces usually access a subkey in digest
+            subkey = n[2].lower()
+            if subkey in p_digest:
+                p_digest = p_digest[subkey]
+
+        if key not in p_digest:
+            if namespace in self.descriptor.namespaces:
+                p_digest = self.descriptor.namespaces[namespace]
+                if key not in p_digest:
+                    raise Exception(f"{key} not present in digest and {namespace}")
+            else:
+                raise Exception(f"{key} not present in digest")
+
+        return key, p_digest[key]
+
+
+    def _handler_default(self, method: str, namespace: str, payload: dict):
+        """
+        This is an euristhic to try parse a namespace carrying state stored in all->digest
+        If the state is not stored in all->digest we'll search our namespace(s) list for
+        state carried through our GETACK messages in the trace
+        """
+        # quick lookup since state is saved in our namespaces
+        if (method == mc.METHOD_GET) and (namespace in self.descriptor.namespaces):
+            return mc.METHOD_GETACK, self.descriptor.namespaces[namespace]
+
+        key, p_state = self._get_key_state(namespace)
+        if method != mc.METHOD_SET:
+            # TODO.....
+            raise Exception(f"{method} not supported in emulator")
+
+        def _update(payload: dict):
+            channel = payload[mc.KEY_CHANNEL]
+            for p in p_state:
+                if p[mc.KEY_CHANNEL] == channel:
+                    p.update(payload)
+                    break
+            else:
+                raise Exception(f"{channel} not present in digest.{key}")
+
+        p_payload = payload[key]
+        if isinstance(p_state, list):
+            if isinstance(p_payload, list):
+                for p_p in p_payload:
+                    _update(p_p)
+            else:
+                _update(p_payload)
+        else:
+            if p_state[mc.KEY_CHANNEL] == p_payload[mc.KEY_CHANNEL]:
+                p_state.update(p_payload)
+            else:
+                raise Exception(f"{p_payload[mc.KEY_CHANNEL]} not present in digest.{key}")
+
+        return mc.METHOD_SETACK, {}
+
+
+    def _GET_Appliance_System_DNDMode(self, header, payload):
+        return mc.METHOD_GETACK, self.p_dndmode
+
+
+    def _SET_Appliance_System_DNDMode(self, header, payload):
+        self.p_dndmode = payload
+        return mc.METHOD_SETACK, {}
+
+
+    def _get_control_key(self, key):
+        p_control = self.descriptor.all.get(mc.KEY_CONTROL)
+        if p_control is None:
+            raise Exception(f"{mc.KEY_CONTROL} not present")
+        if key not in p_control:
+            raise Exception(f"{key} not present in control")
+        return p_control[key]
+
+
+    def _GET_Appliance_Control_Toggle(self, header, payload):
+        # only acual example of this usage comes from legacy firmwares
+        # carrying state in all->control
+        return mc.METHOD_GETACK, { mc.KEY_TOGGLE: self._get_control_key(mc.KEY_TOGGLE) }
+
+
+    def _SET_Appliance_Control_Toggle(self, header, payload):
+        # only acual example of this usage comes from legacy firmwares
+        # carrying state in all->control
+        self._get_control_key(mc.KEY_TOGGLE)[mc.KEY_ONOFF] = payload[mc.KEY_TOGGLE][mc.KEY_ONOFF]
+        return mc.METHOD_SETACK, {}
+
+
+    def _SET_Appliance_Control_Light(self, header, payload):
+        # need to override basic handler since lights turning on/off is tricky between
+        # various firmwares: some supports onoff in light payload some use the togglex
+        p_light = payload[mc.KEY_LIGHT]
+        p_digest = self.descriptor.digest
+        support_onoff_in_light = mc.KEY_ONOFF in p_digest[mc.KEY_LIGHT]
+        # generally speaking set_light always turns on, unless the payload carries onoff = 0 and
+        # the device is not using togglex
+        if support_onoff_in_light:
+            onoff = p_light.get(mc.KEY_ONOFF, 1)
+            p_light[mc.KEY_ONOFF] = onoff
+        else:
+            onoff = 1
+            p_light.pop(mc.KEY_ONOFF, None)
+        if mc.KEY_TOGGLEX in p_digest:
+            # fixed channel 0..that is..
+            p_digest[mc.KEY_TOGGLEX][0][mc.KEY_ONOFF] = onoff
+        p_digest[mc.KEY_LIGHT].update(p_light)
+        return mc.METHOD_SETACK, {}
+
+
+    def _SET_Appliance_Control_Mp3(self, header, payload):
+        if mc.NS_APPLIANCE_CONTROL_MP3 not in self.descriptor.namespaces:
+            raise Exception(f"{mc.NS_APPLIANCE_CONTROL_MP3} not supported in namespaces")
+        mp3 = self.descriptor.namespaces[mc.NS_APPLIANCE_CONTROL_MP3]
+        mp3[mc.KEY_MP3].update(payload[mc.KEY_MP3])
+        return mc.METHOD_SETACK, {}
